requirements.txt
colorama==0.4.6
config==0.5.1
websockets==15.0.1
irc==20.5.0
peewee==3.17.9

## The following requirements were added by pip freeze:
autocommand==2.2.2
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.logging==3.3.0
jaraco.stream==3.0.4
jaraco.text==4.0.0
more-itertools==10.7.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tempora==5.8.0
google-generativeai==0.8.5
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
fastmcp==2.2.5
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google_search_results==2.4.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
markdown-it-py==3.0.0
mcp==1.6.0
mcp-flight-search==0.2.1
mdurl==0.1.2
openapi-pydantic==0.5.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
python-dotenv==1.1.0
requests==2.32.3
rich==14.0.0
rsa==4.9.1
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==2.3.3
starlette==0.46.2
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
google-genai